moved user change into correct namespace
