+++ Portal/src/Datahub.Core/Model/Context/DatahubProjectDBContext.cs
@@ -1,5 +1,4 @@
 using Datahub.Core.Data;
-using Datahub.Core.Model.Achievements;
 using Datahub.Core.Model.Announcements;
 using Datahub.Core.Model.Catalog;
 using Datahub.Core.Model.CloudStorage;
